cast to int to avoid compilation issue with G4 10.06

+++ source/BeastMagnetDetector.cc
@@ -93,7 +93,9 @@ void BeastMagnetDetector::InsertVolumes(G4VPhysicalVolume *physvol)
   G4LogicalVolume *logvol = physvol->GetLogicalVolume();
   m_DisplayAction->AddLogicalVolume(logvol);
   m_PhysicalVolumesSet.insert(physvol);
-  for (int i = 0; i < logvol->GetNoDaughters(); ++i)
+// G4 10.06 returns unsigned int for GetNoDaughters()
+// lower version return int, need to cast to avoid compiler error
+  for (int i = 0; i < (int) logvol->GetNoDaughters(); ++i)
   {
     G4VPhysicalVolume *physvol = logvol->GetDaughter(i);
     // here we decide which volumes are active
